Archive supports Boost .40. Some 64bit casts added.
git-svn-id: svn://svn.geekisp.com/halite@780 362eace2-830a-0410-8e1a-3f32804a3aca

+++ trunk/src/Halite.hpp
@@ -118,7 +118,7 @@ public:
 	bool logToFile() { return logToFile_; }
 	bool logDebug() { return logDebug_; }
 	const std::wstring& dll() { return dll_; }
-	const int logListLen() { return logListLen_; }
+	const int logListLen() { return static_cast<int>(logListLen_); }
 	bool showMessage() { return showMessage_; }
 	
 	friend class GeneralOptions;
